refactor: get data at boot
