+++ src/components/Pages/Boot.tsx
@@ -1,12 +1,19 @@
+/* eslint-disable react-hooks/exhaustive-deps */
 'use client';
 import Image from 'next/image';
 import Apple from '@/assets/icons/apple.svg';
 import { useState, useEffect } from 'react';
 import { useStore } from '@/lib/store';
+import useClock from '@/hooks/useClock';
+import getFullFormatDate from '@/lib/Date/getFullFormatDate';
+import axios from 'axios';
+import { WeatherData } from '@/data/Weather/WeatherData';
 
 export default function Boot() {
   const [progress, setProgress] = useState<number>(0);
-  const { setBooted } = useStore();
+  const [longitude, setLongitude] = useState<number>(59.3326);
+  const [latitude, setLatitude] = useState<number>(18.0649);
+  const { setBooted, setNameOfTheDay, setWeather } = useStore();
 
   useEffect(() => {
     if (progress < 100) {
@@ -21,6 +28,70 @@ export default function Boot() {
       }, 1000);
     }
   }, [progress, setBooted]);
+
+  const currentTime = useClock();
+  const forrmatedTime = getFullFormatDate(currentTime);
+
+  const currentDate =
+    currentTime.getFullYear() +
+    '/' +
+    (currentTime.getMonth() + 1) +
+    '/' +
+    forrmatedTime.dayOfMonth;
+
+  const apiUrl = `http://sholiday.faboul.se/dagar/v2.1/${currentDate}`;
+
+  useEffect(() => {
+    axios
+      .get(apiUrl)
+      .then((response) => {
+        if (
+          response.data &&
+          response.data.dagar &&
+          response.data.dagar.length > 0
+        ) {
+          const namesForToday = response.data.dagar[0].namnsdag;
+          console.log(namesForToday);
+          setNameOfTheDay(namesForToday);
+        }
+      })
+      .catch((error) => {
+        console.error('Error fetching data:', error);
+      });
+  }, []);
+
+  useEffect(() => {
+    const fetchWeatherData = async () => {
+      try {
+        const apiKey = process.env.NEXT_PUBLIC_OPEN_WEATHER;
+        const response = await axios.get(
+          `http://api.openweathermap.org/data/2.5/forecast?lat=${latitude}&lon=${longitude}&appid=${apiKey}&units=metric`
+        );
+        const weatherData: WeatherData = response.data;
+        setWeather(weatherData);
+        console.log(response.data);
+      } catch (err) {
+        console.error(err);
+      }
+    };
+
+    if (navigator.geolocation) {
+      navigator.geolocation.getCurrentPosition(
+        (position) => {
+          const { latitude, longitude } = position.coords;
+          console.log(latitude, longitude);
+          setLatitude(latitude);
+          setLongitude(longitude);
+          fetchWeatherData();
+        },
+        (error) => {
+          console.error(error);
+        }
+      );
+    } else {
+      fetchWeatherData();
+    }
+  }, []);
 
   return (
     <div className="min-h-screen w-full flex flex-col items-center bg-black justify-center">
